% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbands.R
\name{get_bbands}
\alias{get_bbands}
\title{Get Bollinger bands
Calculates upper and lower Bollinger bands based on standard deviation
of stock's close prirce over the past 20 days}
\usage{
get_bbands(stock_ticker)
}
\arguments{
\item{stock_ticker}{String. Stock ticker of company to be analysed}
}
\value{
tibble containing upper and lower Bollinger bands
adjusted closing price, moving average, date
}
\description{
Get Bollinger bands
Calculates upper and lower Bollinger bands based on standard deviation
of stock's close prirce over the past 20 days
}
\examples{
get_bbands("MSFT")
}
