% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Calculate the moving average of a stock price for a given range of days}
\usage{
moving_average(stock_ticker, period)
}
\arguments{
\item{stock_ticker}{A string corresponding to the name of the stock.}

\item{period}{Number of days for which the moving average is computed}
}
\value{
moving_avg A tibble representing the moving average of a stock
price over the specified period.
}
\description{
Calculate the moving average of a stock price for a given range of days
}
\examples{
moving_average("MSFT", 20)
}
