% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_200ma.R
\name{plot_200ma}
\alias{plot_200ma}
\title{Plot stock price and corresponding 200 day moving average.}
\usage{
plot_200ma(stock_ticker)
}
\arguments{
\item{stock_ticker}{string
name of the stock for which the plot is created}
}
\value{
A line chart showing price data and corresponding 200 day
moving average line for a stock.
}
\description{
Plot stock price and corresponding 200 day moving average.
}
\examples{
plot_200ma("MSFT")
}
