% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bbands.R
\name{plot_bbands}
\alias{plot_bbands}
\title{Plot stock price and corresponding upper and lower Bollinger bands
for the last 200 trading days.}
\usage{
plot_bbands(stock_ticker)
}
\arguments{
\item{stock_ticker}{string Ticker symbol of the
stock for which the plot is created}
}
\value{
A line chart showing price data and corresponding upper
and lower Bollinger bands.
}
\description{
Plot stock price and corresponding upper and lower Bollinger bands
for the last 200 trading days.
}
\examples{
plot_bbands("MSFT")
}
