% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ma_10_20days.R
\name{plot_ma_10_20days}
\alias{plot_ma_10_20days}
\title{Plot stock price and corresponding 10 and 20 day moving average.}
\usage{
plot_ma_10_20days(stock_symbol)
}
\arguments{
\item{stock_symbol}{Ticker symbol of the stock for which the plot is created}
}
\value{
A line chart showing price data and corresponding 10 and 20 day moving average line for a stock.
}
\description{
Plot stock price and corresponding 10 and 20 day moving average.
}
\examples{
plot_ma_10_20days("MSFT")

}
