% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFT.R
\name{"MSFT"}
\alias{"MSFT"}
\title{Microsoft data}
\format{
\subsection{\code{MSFT}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<finance.yahoo.com>
}
\description{
Data downloaded on the Microsoft stock
Report ...
}
